requirements.txt
django==4.2.8
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
django-theme-material-kit
django-dynamic-api==1.0.4
fontawesomefree>=6.6.0
pycountry
cryptography>=44.0.2
reportlab>=4.3.1
reportlab-qrcode
django-apscheduler
redis
django-cors-headers
django-admin-material-dashboard
django_countries
channels
channels-redis
asgiref
pillow
faker
langdetect
requests
djangorestframework
netifaces
